requirements.txt
Flask==2.2.2
future==0.18.2
numpy==1.23.4
opencv_python==4.6.0.66
Pillow==9.4.0
